led in main aber crash

+++ main/pulselight.h
@@ -7,8 +7,8 @@ class Pulselight {
   public:
     void LED_setup();
     void LED_end();
-    float LED_setfq(float fq);
-    void LED_update(float m);
+    double LED_setfq(double fq);
+    void LED_update(double m);
 };
 
 #endif
